feat: use typeorm for board model
